refactor: 주석 코드 제거 및 패키지 구조 수정
