+++ src/main/java/capstone/project/credit_manager/config/auth/JwtAuthenticationFilter.java
@@ -35,16 +35,6 @@ public class JwtAuthenticationFilter extends GenericFilterBean {
         }
     }
 
-//    @Override
-//    public void doFilter(ServletRequest request, ServletResponse response, FilterChain filterChain) throws IOException, ServletException {
-//        String token = jwtTokenProvider.resolveToken((HttpServletRequest) request);
-//        if (token != null && jwtTokenProvider.validateToken(token)) {
-//            Authentication auth = jwtTokenProvider.getAuthentication(token);
-//            SecurityContextHolder.getContext().setAuthentication(auth);
-//        }
-//        filterChain.doFilter(request, response);
-//    }
-
     private void InvalidTokenExceptionHandler(HttpServletResponse response) throws IOException {
         ErrorResponse error = new ErrorResponse("UNAUTHORIZED", "유효하지 않은 인증토큰 입니다.");
         response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
