+++ src/main/java/capstone/project/credit_manager/web/controller/AccountRestApiController.java
@@ -4,6 +4,7 @@ import capstone.project.credit_manager.config.auth.JwtTokenProvider;
 import capstone.project.credit_manager.domain.accounts.Account;
 import capstone.project.credit_manager.service.AccountService;
 import capstone.project.credit_manager.service.LoginInfoRequestDto;
+import capstone.project.credit_manager.web.controller.vo.JwtAccessToken;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
